Use pundit and tests

+++ app/controllers/skills_controller.rb
@@ -30,6 +30,7 @@ class SkillsController < ApplicationController
   end
 
   def update
+    authorize @skill
     if @skill.update skill_params
       flash[:notice] = 'Successfully updated'
     else
@@ -39,12 +40,15 @@ class SkillsController < ApplicationController
   end
 
   def create
-    Skill.create skill_params
+    @skill = Skill.new(skill_params)
+    authorize @skill
+    @skill.save
     flash[:notice] = 'Successfully created'
     redirect_to skills_path
   end
 
   def destroy
+    authorize @skill
     if @skill.destroy
       flash[:notice] = 'Successfully destroyed'
     else
@@ -61,5 +65,9 @@ class SkillsController < ApplicationController
 
   def skill_params
     params.require(:skill).permit!
+  end
+
+  def pundit_user
+    current_employee
   end
 end
